added footer structure
